kontrol_s5: initial commit, see #98
This commit imports a whole bunch of code from the Maschine MK3
driver, and reworks it for the Kontrol S5. Most of the code here
is garbage, and needs to be removed / cleaned up.

When running examples/ctlra_simple, the driver now turns on all
LEDs, and on shutdown Ctrl+C, it turns them off again.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>